alarm clock
This project is all about design of a alarm clock to display time in LCD in a 24hr format. Developed the RTL code for each of the sub-blocks used in the block level architecture of the Alarm clock and verified each sub-block using task-based Verilog Testbench. Finally verified the top-level RTL using Verilog Testbench